# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/mapper/ItemMapper.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/entity/SharedFlat.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/DigitalStorageServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/ItemServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/service/impl/SharedFlatService.java
#	frontend/package-lock.json
#	frontend/package.json
#	frontend/src/app/app-routing.module.ts
#	frontend/src/app/app.module.ts

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/endpoint/dto/UnitConvertDto.java
@@ -6,6 +6,6 @@ import io.soabase.recordbuilder.core.RecordBuilder;
 public record UnitConvertDto(
     UnitDto from,
     UnitDto to,
-    Long value
+    Double value
 ) {
 }
